test: move existing COPY format parsing tests into a single module
These tests are independent from sync/async code.
